MNT: wrap frompyfunc in Python
