+++ numpy/_core/umath.py
@@ -7,6 +7,7 @@ by importing from the extension module.
 """
 
 import numpy
+from numpy._utils import set_module
 
 from . import _multiarray_umath
 from ._multiarray_umath import *
@@ -21,6 +22,7 @@ from ._multiarray_umath import (
     _expandtabs,
     _expandtabs_length,
     _extobj_contextvar,
+    _frompyfunc,
     _get_extobj_dict,
     _ljust,
     _lstrip_chars,
@@ -57,3 +59,56 @@ __all__ = [
     'power', 'rad2deg', 'radians', 'reciprocal', 'remainder', 'right_shift',
     'rint', 'sign', 'signbit', 'sin', 'sinh', 'spacing', 'sqrt', 'square',
     'subtract', 'tan', 'tanh', 'true_divide', 'trunc', 'vecdot', 'vecmat']
+
+
+@set_module("numpy")
+def frompyfunc(func, /, nin, nout, **kwargs):
+    """
+
+    Takes an arbitrary Python function and returns a NumPy ufunc.
+
+    Can be used, for example, to add broadcasting to a built-in Python
+    function (see Examples section).
+
+    Parameters
+    ----------
+    func : Python function object
+        An arbitrary Python function.
+    nin : int, optional
+        The number of input arguments. Defaults to `func.__code__.co_argcount`.
+    nout : int, optional
+        The number of objects returned by `func`. Defaults to 1.
+    identity : object, optional
+        The value to use for the `~numpy.ufunc.identity` attribute of the resulting
+        object. If specified, this is equivalent to setting the underlying
+        C ``identity`` field to ``PyUFunc_IdentityValue``.
+        If omitted, the identity is set to ``PyUFunc_None``. Note that this is
+        _not_ equivalent to setting the identity to ``None``, which implies the
+        operation is reorderable.
+
+    Returns
+    -------
+    out : ufunc
+        Returns a NumPy universal function (``ufunc``) object.
+
+    See Also
+    --------
+    vectorize : Evaluates pyfunc over input arrays using broadcasting rules of numpy.
+
+    Notes
+    -----
+    The returned ufunc always returns PyObject arrays.
+
+    Examples
+    --------
+    Use frompyfunc to add broadcasting to the Python function ``oct``:
+
+    >>> import numpy as np
+    >>> oct_array = np.frompyfunc(oct, 1, 1)
+    >>> oct_array(np.array((10, 30, 100)))
+    array(['0o12', '0o36', '0o144'], dtype=object)
+    >>> np.array((oct(10), oct(30), oct(100))) # for comparison
+    array(['0o12', '0o36', '0o144'], dtype='<U5')
+
+    """
+    return _frompyfunc(func, nin=nin, nout=nout, **kwargs)
